Move Utils.describe() to Blocks.describe()
