+++ api/src/main/java/ai/djl/util/Utils.java
@@ -13,9 +13,6 @@
 package ai.djl.util;
 
 import ai.djl.ndarray.NDArray;
-import ai.djl.ndarray.types.Shape;
-import ai.djl.nn.Block;
-import ai.djl.nn.LambdaBlock;
 import ai.djl.nn.Parameter;
 
 import org.slf4j.Logger;
@@ -39,7 +36,6 @@ import java.util.Scanner;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 /** A class containing utility methods. */
 public final class Utils {
@@ -356,52 +352,5 @@ public final class Utils {
             }
         }
         return Paths.get(cacheDir);
-    }
-
-    /**
-     * Returns a string representation of the passed {@link Block} describing the input axes, output
-     * axes, and the block's children.
-     *
-     * @param block the block to describe
-     * @param blockName the name to be used for the passed block, or <code>null</code> if its class
-     *     name is to be used
-     * @param beginAxis skips all axes before this axis; use <code>0</code> to print all axes and
-     *     <code>1</code> to skip the batch axis.
-     * @return the string representation
-     */
-    public static String describe(Block block, String blockName, int beginAxis) {
-        Shape[] inputShapes = block.isInitialized() ? block.getInputShapes() : null;
-        Shape[] outputShapes = inputShapes != null ? block.getOutputShapes(inputShapes) : null;
-        StringBuilder sb = new StringBuilder(200);
-        if (block instanceof LambdaBlock
-                && !LambdaBlock.DEFAULT_NAME.equals(((LambdaBlock) block).getName())) {
-            sb.append(((LambdaBlock) block).getName());
-        } else if (blockName != null) {
-            sb.append(blockName);
-        } else {
-            sb.append(block.getClass().getSimpleName());
-        }
-        if (inputShapes != null) {
-            sb.append(
-                    Stream.of(inputShapes)
-                            .map(shape -> shape.slice(beginAxis).toString())
-                            .collect(Collectors.joining("+")));
-        }
-        if (!block.getChildren().isEmpty()) {
-            sb.append(" {\n");
-            for (Pair<String, Block> pair : block.getChildren()) {
-                String child = describe(pair.getValue(), pair.getKey().substring(2), beginAxis);
-                sb.append(child.replaceAll("(?m)^", "\t")).append('\n');
-            }
-            sb.append('}');
-        }
-        if (outputShapes != null) {
-            sb.append(" -> ");
-            sb.append(
-                    Stream.of(outputShapes)
-                            .map(shape -> shape.slice(beginAxis).toString())
-                            .collect(Collectors.joining("+")));
-        }
-        return sb.toString();
     }
 }
